feat(outbox): enhance outbox management with new admin features
- Introduced OutboxAdminController and OutboxAdminService for managing dead-lettered outbox rows.
- Implemented endpoints for listing, requeuing single and bulk dead-lettered rows, and checking relay health.
- Added new permissions for outbox management in the RBAC policy.
- Expanded integration event payloads to include MarketTickerUpdatedOutboxPayload.
- Updated tests to cover new outbox management functionalities and ensure operational integrity.

+++ src/entities/currency.entity.ts
@@ -15,7 +15,7 @@ export class Currency {
   @Column({ type: 'varchar', length: 64 })
   name!: string;
 
-  @Column({ type: 'tinyint', default: 8 })
+  @Column({ type: 'smallint', default: 8 })
   precision_scale!: number;
 
   @Column({ ...DECIMAL_36_18_DEFAULT_0_COLUMN })
